[qtcontacts-sqlite] Compile with Qt5.6. Contributes to JB#35409

Do not merge until Qt 5.6 has merged.

See merge request !5

+++ src/extensions/qcontactstatusflags_impl.h
@@ -55,7 +55,7 @@ void QContactStatusFlags::setFlags(Flags flags)
 
 QContactStatusFlags::Flags QContactStatusFlags::flags() const
 {
-    return Flags(flagsValue());
+    return Flags(static_cast<ulong>(flagsValue()));
 }
 
 void QContactStatusFlags::setFlagsValue(quint64 value)
